pyproject.toml
[tool.poetry]
name = "nextassist"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
snowflake-core = "^1.0.2"
snowflake-snowpark-python = "^1.26.0"
snowflake-ml-python = "^1.7.2"
llama-index-core = "^0.12.10.post1"
trulens = "^1.3.0"
trulens-connectors-snowflake = "^1.3.2"
trulens-providers-cortex = "^1.3.2"
json-repair = "^0.35.0"
trulens-providers-openai = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
